refactor: modify the database to have user_id

+++ src/models/links-model.js
@@ -3,32 +3,34 @@ import { pool } from "../db/dbConnect.js";
 export async function create({
     code,
     target_url,
-    expires_at
+    expires_at,
+    user_id
 }) {
     const { rows } = await pool.query(
         `
         INSERT INTO links
-        ( code, target_url, expires_at ) VALUES
-        ( $1, $2, $3
+        ( code, target_url, expires_at , user_id ) VALUES
+        ( $1, $2, $3, $4
         )
         RETURNING *
         `,
-        [ code, target_url, expires_at ?? null ]
+        [ code, target_url, expires_at ?? null , user_id]
     );
 
     return rows[0];
 }
 
 
-export async function findLinkByCode(code) {
+export async function findLinkByCode(code, user_id) {
   const { rows } = await pool.query(
     `
     SELECT * 
     FROM links 
     WHERE code = $1
+    AND user_id = $2
 
     `, 
-      [code]
+      [code, user_id]
   );
 
   return rows[0];
@@ -45,8 +47,7 @@ export async function clickCount(linkId, referrer, userAgent) {
         UPDATE links
         SET click_count = click_count + 1
         WHERE id = $1
-        
-
+      
         `,
           [linkId]
       );
@@ -75,7 +76,7 @@ export async function clickCount(linkId, referrer, userAgent) {
 }
 
 
-export async function getMetadata(code) {
+export async function getMetadata(code, user_id) {
 
   const { rows } = await pool.query(
     `
@@ -83,30 +84,32 @@ export async function getMetadata(code) {
       code, target_url, click_count, created_at, expires_at
     FROM links
     WHERE code = $1
+    AND user_id = $2
     
 
-    `, [code]
+    `, [code, user_id]
   );
 
   return rows[0];
 }
 
-export async function remove(code) {
+export async function remove(code, user_id) {
 
   const result = await pool.query(
     `
     DELETE FROM links
     WHERE code = $1
+    AND user_id = $2
     
 
-    `, [code]
+    `, [code, user_id]
   );
 
   return result.rowCount
 }
 
 
-export async function getLinkClicks (code, after = 0, limit = 10) {
+export async function getLinkClicks (code, user_id, after = 0, limit = 10) {
 
   const { rows } = await pool.query(
     `
@@ -115,7 +118,7 @@ export async function getLinkClicks (code, after = 0, limit = 10) {
     FROM clicks c
     JOIN links l
       ON c.link_id = l.id
-    WHERE l.code = $1 AND c.id > $2
+    WHERE l.code = $1 AND l.user_id = $2 AND c.id > $3
     ORDER BY c.id
     LIMIT $3
